Configure: update to latest autoconf strategy
In p4est, we had removed the @P4EST_SC_AMFLAGS@ variable from Makefile.am
since autoreconf would die on it.  But we have found that this causes
crashes of configure on rebuilding, since the -i flags supplied to
autoreconf apparently do not make it through into the build.

So we revert to not using autoreconf in p4est, which parallels t8code.
We bump p4est and sc to their latest versions and remove the m4 cache.